// packages/core/src/types.ts
export class ExitButtonError extends Error {
  code: string;
  statusCode?: number;

  constructor(message: string, code: string, statusCode?: number) {
    super(message);
    this.name = 'ExitButtonError';
    this.code = code;
    this.statusCode = statusCode;
  }
}